Financial program.
Finie is a program that is made to help one bank and use money properly. i made this project to showcase my basic understanding of Java.
This project will be improved and made better with every new thing i learn about coding. Still in its beginner phase and i hope when i Graduate (in 3 years) it will be advanced and proper.